refactor(#792): isolate chain_row_wiring tests into their own file (sweep)
chain_row_wiring.rs 795 -> 621. The inline tests move to a sibling wired via
#[path]; production body still exceeds the cap so it stays on the debt list,
but the test bulk is now separated. adapter-gui lib tests (537) pass, zero warnings.
